# docs/swagger.yaml
basePath: /
definitions:
  github_com_SecureParadise_go_attendence_internal_db_sqlc.Student:
    properties:
      batch:
        $ref: '#/definitions/pgtype.Text'
      branch_id:
        type: string
      created_at:
        type: string
      current_semester_id:
        type: string
      deleted_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      fingerprint_hash:
        $ref: '#/definitions/pgtype.Text'
      first_name:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/pgtype.Text'
      last_name:
        type: string
      middle_name:
        $ref: '#/definitions/pgtype.Text'
      rfid_tag_id:
        $ref: '#/definitions/pgtype.Text'
      roll_no:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_SecureParadise_go_attendence_internal_db_sqlc.Teacher:
    properties:
      card_no:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      department_id:
        type: string
      fingerprint_hash:
        $ref: '#/definitions/pgtype.Text'
      first_name:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/pgtype.Text'
      last_name:
        type: string
      middle_name:
        $ref: '#/definitions/pgtype.Text'
      rfid_tag_id:
        $ref: '#/definitions/pgtype.Text'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_SecureParadise_go_attendence_internal_db_sqlc.User:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      department_id:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_email_verified:
        type: boolean
      is_profile_completed:
        type: boolean
      last_login_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      password_changed_at:
        $ref: '#/definitions/pgtype.Timestamptz'
      password_hash:
        type: string
      updated_at:
        type: string
      user_role:
        $ref: '#/definitions/github_com_SecureParadise_go_attendence_internal_db_sqlc.Userrole'
    type: object
  github_com_SecureParadise_go_attendence_internal_db_sqlc.Userrole:
    enum:
    - student
    - teacher
    - hod
    - dhod
    - admin
    - crew
    type: string
    x-enum-varnames:
    - UserroleStudent
    - UserroleTeacher
    - UserroleHod
    - UserroleDhod
    - UserroleAdmin
    - UserroleCrew
  internal_api_handlers.CreateStudentRequest:
    properties:
      batch:
        type: string
      branch_code:
        type: string
      email:
        type: string
      first_name:
        type: string
      image:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      roll_no:
        type: string
      semester_no:
        type: integer
    required:
    - batch
    - branch_code
    - email
    - first_name
    - last_name
    - roll_no
    - semester_no
    type: object
  internal_api_handlers.CreateTeacherRequest:
    properties:
      card_no:
        type: string
      department_name:
        type: string
      email:
        type: string
      first_name:
        type: string
      image:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
    required:
    - card_no
    - department_name
    - email
    - first_name
    - last_name
    type: object
  internal_api_handlers.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      user_role:
        $ref: '#/definitions/github_com_SecureParadise_go_attendence_internal_db_sqlc.Userrole'
    required:
    - email
    - password
    type: object
  internal_api_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_api_handlers.LoginResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      is_profile_completed:
        type: boolean
      role:
        type: string
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a student attendance management system server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Attendance API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, password, and optional role
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - users
  /student/{roll_no}:
    get:
      consumes:
      - application/json
      description: Fetch student details using their roll number
      parameters:
      - description: Roll Number
        in: path
        name: roll_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SecureParadise_go_attendence_internal_db_sqlc.Student'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get student by roll number
      tags:
      - students
  /student_reg:
    post:
      consumes:
      - application/json
      description: Complete student profile with personal and academic details
      parameters:
      - description: Student profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_SecureParadise_go_attendence_internal_db_sqlc.Student'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete student profile
      tags:
      - students
  /teacher/{card_no}:
    get:
      consumes:
      - application/json
      description: Fetch teacher details using their card number
      parameters:
      - description: Card Number
        in: path
        name: card_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SecureParadise_go_attendence_internal_db_sqlc.Teacher'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get teacher by card number
      tags:
      - teachers
  /teacher_reg:
    post:
      consumes:
      - application/json
      description: Complete teacher profile with personal and academic details
      parameters:
      - description: Teacher profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_SecureParadise_go_attendence_internal_db_sqlc.Teacher'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete teacher profile
      tags:
      - teachers
  /user/me:
    get:
      consumes:
      - application/json
      description: Get profile of the user identified by the access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SecureParadise_go_attendence_internal_db_sqlc.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"